+++ SINGER/SINGER/Coalescent_calculator.cpp
@@ -7,7 +7,7 @@
 
 #include "Coalescent_calculator.hpp"
 
-Coalescent_calculator::Coalescent_calculator(float t) {
+Coalescent_calculator::Coalescent_calculator(double t) {
     cut_time = t;
 }
 
@@ -21,17 +21,17 @@ void Coalescent_calculator::compute(set<Branch> &branches) {
     assert(rates.size() == rate_changes.size());
 }
 
-float Coalescent_calculator::weight(float lb, float ub) {
-    float p = prob(ub) - prob(lb);
+double Coalescent_calculator::weight(double lb, double ub) {
+    double p = prob(ub) - prob(lb);
     assert(!isnan(p));
     return p;
 }
 
-float Coalescent_calculator::time(float lb, float ub) {
-    float lq = prob(lb);
-    float uq = prob(ub);
-    float mid = 0;
-    float t;
+double Coalescent_calculator::time(double lb, double ub) {
+    double lq = prob(lb);
+    double uq = prob(ub);
+    double mid = 0;
+    double t;
     if (isinf(ub)) {
         return lb + log(2);
     }
@@ -47,8 +47,8 @@ float Coalescent_calculator::time(float lb, float ub) {
 
 void Coalescent_calculator::compute_rate_changes(set<Branch> &branches) {
     rate_changes.clear();
-    float lb = 0;
-    float ub = 0;
+    double lb = 0;
+    double ub = 0;
     for (Branch b : branches) {
         lb = max(cut_time, b.lower_node->time);
         ub = b.upper_node->time;
@@ -73,11 +73,11 @@ void Coalescent_calculator::compute_probs_quantiles() {
     probs.clear();
     quantiles.clear();
     int curr_rate = 0;
-    float prev_time = 0.0;
-    float next_time = 0.0;
-    float prev_prob = 1.0;
-    float next_prob = 1.0;
-    float cum_prob = 0.0;
+    double prev_time = 0.0;
+    double next_time = 0.0;
+    double prev_prob = 1.0;
+    double next_prob = 1.0;
+    double cum_prob = 0.0;
     for (auto it = rates.begin(); it != prev(rates.end()); it++) {
         curr_rate = it->second;
         prev_time = it->first;
@@ -99,13 +99,13 @@ void Coalescent_calculator::compute_probs_quantiles() {
     assert(quantiles.size() == rates.size());
 }
 
-float Coalescent_calculator::prob(float x) {
+double Coalescent_calculator::prob(double x) {
     if (x > max_time) {
         x = max_time;
     } else if (x < min_time) {
         x = min_time;
     }
-    pair<float, float> query = {x, -1.0f};
+    pair<double, double> query = {x, -1.0f};
     auto it = probs.find(query);
     if (it != probs.end()) {
         return it->second;
@@ -113,31 +113,31 @@ float Coalescent_calculator::prob(float x) {
     auto u_it = probs.upper_bound(query);
     auto l_it = probs.upper_bound(query);
     l_it--;
-    float base_prob = l_it->second;
+    double base_prob = l_it->second;
     int rate = rates[l_it->first];
     if (rate == 0) {
         return base_prob;
     }
-    float delta_t = u_it->first - l_it->first;
-    float delta_p = u_it->second - l_it->second;
-    float new_delta_t = x - l_it->first;
-    // float new_delta_p = delta_p*(1 - exp(-rate*new_delta_t))/(1 - exp(-rate*delta_t));
-    float new_delta_p = delta_p*expm1(-rate*new_delta_t)/expm1(-rate*delta_t);
+    double delta_t = u_it->first - l_it->first;
+    double delta_p = u_it->second - l_it->second;
+    double new_delta_t = x - l_it->first;
+    // double new_delta_p = delta_p*(1 - exp(-rate*new_delta_t))/(1 - exp(-rate*delta_t));
+    double new_delta_p = delta_p*expm1(-rate*new_delta_t)/expm1(-rate*delta_t);
     assert(!isnan(new_delta_p));
     return base_prob + new_delta_p;
 }
 
-float Coalescent_calculator::quantile(float p) {
-    pair<float, float> query = {-1.0f, p};
+double Coalescent_calculator::quantile(double p) {
+    pair<double, double> query = {-1.0f, p};
     auto u_it = quantiles.upper_bound(query);
     auto l_it = quantiles.upper_bound(query);
     l_it--;
-    float base_time = l_it->first;
+    double base_time = l_it->first;
     int rate = rates[l_it->first];
-    float delta_t = u_it->first - l_it->first;
-    float delta_p = u_it->second - l_it->second;
-    float new_delta_p = p - l_it->second;
-    float new_delta_t = 1 - new_delta_p/delta_p*(1 - exp(-rate*delta_t));
+    double delta_t = u_it->first - l_it->first;
+    double delta_p = u_it->second - l_it->second;
+    double new_delta_p = p - l_it->second;
+    double new_delta_t = 1 - new_delta_p/delta_p*(1 - exp(-rate*delta_t));
     new_delta_t = -log(new_delta_t)/rate;
     assert(!isnan(new_delta_t));
     return base_time + new_delta_t;
