+++ SINGER/SINGER/Interval.hpp
@@ -20,30 +20,30 @@ class Interval {
 public:
     
     Branch branch = Branch();
-    float lb = 0;
-    float ub = 0;
-    float weight = 0.0;
-    float time = 0.0;
+    double lb = 0;
+    double ub = 0;
+    double weight = 0.0;
+    double time = 0.0;
     int start_pos = 0;
     int source_pos = 0;
     Node_ptr node = nullptr;
-    float reduction = 1.0;
+    double reduction = 1.0;
     
-    vector<float> source_weights = {};
+    vector<double> source_weights = {};
     vector<Interval *> source_intervals = {};
     vector<shared_ptr<Interval>> intervals = {};
     
     Interval();
     
-    Interval(Branch b, float tl, float tu, int init_pos);
+    Interval(Branch b, double tl, double tu, int init_pos);
     
-    void assign_weight(float w);
+    void assign_weight(double w);
     
-    void assign_time(float t);
+    void assign_time(double t);
     
     void fill_time();
     
-    bool full(float t);
+    bool full(double t);
 
     bool operator<(const Interval &other) const;
     
@@ -52,7 +52,7 @@ public:
     bool operator!=(const Interval &other) const;
 };
 
-shared_ptr<Interval> create_interval(Branch b, float tl, float tu, int init_pos);
+shared_ptr<Interval> create_interval(Branch b, double tl, double tu, int init_pos);
 
 struct compare_interval {
     
@@ -86,14 +86,14 @@ class Interval_info {
 public:
     
     Branch branch;
-    float lb = 0;
-    float ub = 0;
-    float time = 0;
-    float seed_pos = 0;
+    double lb = 0;
+    double ub = 0;
+    double time = 0;
+    double seed_pos = 0;
     
     Interval_info();
     
-    Interval_info(Branch b, float tl, float tu);
+    Interval_info(Branch b, double tl, double tu);
     
     bool operator==(const Interval_info& other) const;
